+++ packages/compass-connections/src/index.ts
@@ -1,4 +1,4 @@
-import SingleConnectionForm from './components/connections';
+import SingleConnectionForm from './components/legacy-connections';
 import { ConnectionsProvider } from './components/connections-provider';
 export { LegacyConnectionsModal } from './components/legacy-connections-modal';
 export { SingleConnectionForm };
